# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "🐐 基于 gin + gorm 的轻量级工程项目",
        "title": "接口文档",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1beta1/book": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "获取书籍列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1beta1/example": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Get请求",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1beta1/example/err": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Err请求",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1beta1/example/form": {
            "post": {
                "description": "处理application/x-www-form-urlencoded类型的POST请求",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "form表单请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1beta1/example/json": {
            "post": {
                "description": "邮箱、用户名校验",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "JSON参数请求",
                "parameters": [
                    {
                        "description": "测试请求json参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1beta1.JSONBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1beta1/example/query": {
            "get": {
                "description": "查询参数，匹配 query?id=xxx",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "query参数查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1beta1/example/uri/{id}": {
            "get": {
                "description": "路径参数，匹配 /uri/{id}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "uri参数请求",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id值",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "v1beta1.JSONBody": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@qq.com"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}